pyproject.toml
[tool.poetry]
name = "very-fire-skills"
description = "Very Fire"
version = "0.1.0"
readme = "README.md"
authors = []
classifiers = [
  "Private :: Do not Upload",
]


[tool.poetry.dependencies]
python = "3.12.1"

django = { version = "^4.2", extras = ["argon2"] }
django-split-settings = "^1.3"
django-axes = "^6.4"
django-csp = "^3.8"
django-health-check = "^3.18"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.19"
django-stubs-ext = "^5.0"

psycopg2-binary = "^2.9"
gunicorn = "^22.0"
python-decouple = "^3.8"
structlog = "^24.1"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
model-bakery = "^1.18.1"
gigachat = "^0.1.31"
yandexgptlite = "^0.1.7"
ollama = "^0.2.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3"
django-migration-linter = "^5.1"
django-extra-checks = "^0.14"
nplusone = "^1.0"

wemake-python-styleguide = "^0.18"
flake8-pytest-style = "^2.0"
flake8-django = "^1.4"
flake8-logging-format = "^0.9"
nitpick = "^0.35"
doc8 = "^1.0"

pytest = "^8.2.1"
pytest-django = "^4.8"
pytest-cov = "^5.0"
covdefaults = "^2.3"
django-coverage-plugin = "^3.1"
pytest-randomly = "^3.15"
pytest-timeout = "^2.3"
django-test-migrations = "^1.3"
hypothesis = "^6.102"

django-stubs = { version = "^4.2.7", extras = ["compatible-mypy"] }

djlint = "^1.34"
yamllint = "^1.35"
safety = "^3.2"
dotenv-linter = "^0.5"
polint = "^0.4"
dennis = "^1.1"
dump-env = "^1.5"
ipython = "^8.24"
django-query-counter = "^0.4"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.3"
sphinx-autodoc-typehints = "^2.1"
tomli = "^2.0"


[build-system]
requires = ["poetry-core>=1.9"]
build-backend = "poetry.core.masonry.api"


[tool.djlint]
ignore = "H006,H030,H031"
include = "H017,H035"
indent = 2
blank_line_after_tag = "load,extends"
profile = "django"
max_line_length = 80
format_attribute_template_tags = true


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"
